+++ R/evaluate_imputed_values.R
@@ -24,6 +24,8 @@
 #' of the true values.
 #' These relative versions can be selected via "bias_rel" and "MAE_rel".
 #'
+#' The argument \code{which_cols} allows the selection of columns
+#' for comparison (see examples).
 #'
 #' If \code{M = NULL} (the default), then all values of \code{imp_ds} and
 #' \code{orig_ds} will be used for the calculation of the evaluation criterion.
@@ -34,11 +36,13 @@
 #' same dimensions as \code{orig_ds} and missing values must be coded as TRUE.
 #' This is the standard behavior, if you use \code{\link[base]{is.na}} on a
 #' dataset with missing values to generate \code{M} (see examples).
+#' It is possible to combine \code{M} and \code{which_cols}.
 #'
 #' @param imp_ds a data frame or matrix with imputed values
 #' @param orig_ds a data frame or matrix with original (true) values
 #' @param criterion a string specifying the used criterion for comparing the
 #'   imputed and original values
+#' @param which_cols indices or names of columns used for evaluation
 #' @param M NULL (the default) or a missing data indicator matrix; the missing
 #'   data indicator matrix is normally created via \code{is.na(miss_ds)}, where
 #'   \code{miss_ds} is the dataset after deleting values from \code{orig_ds}
@@ -54,10 +58,19 @@
 #' M <- is.na(miss_ds)
 #' imp_ds <- impute_mean(miss_ds)
 #' evaluate_imputed_values(imp_ds, orig_ds, M = M)
+#' # compare only the imputed values in column X
+#' evaluate_imputed_values(imp_ds, orig_ds, M = M, which_cols = "X")
 evaluate_imputed_values <- function(imp_ds, orig_ds, criterion = "RMSE", M = NULL,
+                                    which_cols = seq_len(ncol(imp_ds)),
                                     tolerance = sqrt(.Machine$double.eps)) {
   if (!isTRUE(all.equal(dim(imp_ds), dim(orig_ds)))) {
     stop("the dimensions of imp_ds and orig_ds must be equal")
+  }
+
+  imp_ds <- imp_ds[, which_cols, drop = FALSE]
+  orig_ds <- orig_ds[, which_cols, drop = FALSE]
+  if (!is.null(M)) {
+    M <- M[, which_cols, drop = FALSE]
   }
 
   calc_evaluation_criterion(imp_ds, orig_ds, criterion, M, tolerance = tolerance)