+++ NEWS.md
@@ -16,6 +16,7 @@ Functions for evaluation:
 
 * `delete_MAR_1_to_x()` and `delete_MNAR_1_to_x()` can now handle (unordered) factors
 * new criteria for `evaluate_imputed_values()` and `evaluate_parameters()`: three forms of NRMSE, nr_equal, nr_NA and precision
+* `evaluate_imputed_values()`: add argument `which_cols` to select columns for evaluation. 
 
 ## Miscellaneous